refactor: 패키지 이동
DTO 클래스는 Controller에서만 사용하므로 Controller 패키지 내부에서 관리한다.
