+++ src/main/java/com/aid/aidbackend/controller/dto/MemberDto.java
@@ -1,4 +1,4 @@
-package com.aid.aidbackend.dto;
+package com.aid.aidbackend.controller.dto;
 
 
 import jakarta.validation.constraints.Email;
